remove exemplar class
This class was just a wrapper around `Instance`, and had no clear
purpose while also making the code unclear, so remove it.

Notice the strange required use of `toIntValue` in
`AnalogicalModelingOutput`. Weka uses ints in some places and doubles
in others for indexing attribute values. How annoying is that! This
may be a misunderstanding on my part, though.

+++ Analogical Modeling/src/main/weka/classifiers/lazy/AM/lattice/distributed/DistributedLattice.java
@@ -223,7 +223,7 @@ public class DistributedLattice {
 			set.add(i);
 
 		Set<Integer> intersection = new HashSet<Integer>();
-		int outcome = 0;
+		double outcome = 0;
 		for (Integer i : larger)
 			// determine heterogeneity whenever we add a new subcontext
 			if (set.contains(i)) {
